requirements.txt
Flask
flask_cors
nltk
spacy
transformers
requests
python-dotenv
torch